shl-recommender/requirements.txt
flask
faiss-cpu
ollama
bs4
requests
numpy